refactor: extract copy helper
