+++ postbuild.js
@@ -1,23 +1,10 @@
 const fs = require('fs');
 const path = require('path');
+const { copyRecursiveSync } = require('./scripts/copyRecursive');
 
 const folders = ['html', 'css', 'fonts', 'icons'];
 const appDir = path.join(__dirname, 'app');
 const distDir = path.join(__dirname, 'dist', 'app');
-
-function copyRecursiveSync(src, dest) {
-  const stat = fs.statSync(src);
-  if (stat.isDirectory()) {
-    fs.mkdirSync(dest, { recursive: true });
-    for (const entry of fs.readdirSync(src)) {
-      const srcPath = path.join(src, entry);
-      const destPath = path.join(dest, entry);
-      copyRecursiveSync(srcPath, destPath);
-    }
-  } else {
-    fs.copyFileSync(src, dest);
-  }
-}
 
 for (const folder of folders) {
   const src = path.join(appDir, folder);
