+++ watch-assets.js
@@ -1,25 +1,11 @@
 const fs = require('fs');
 const path = require('path');
 const watchboy = require('watchboy');
+const { copyRecursiveSync } = require('./scripts/copyRecursive');
 
 const folders = ['html', 'css', 'fonts', 'icons'];
 const appDir = path.join(__dirname, 'app');
 const distDir = path.join(__dirname, 'dist', 'app');
-
-function copyRecursiveSync(src, dest) {
-  const stat = fs.statSync(src);
-  if (stat.isDirectory()) {
-    fs.mkdirSync(dest, { recursive: true });
-    for (const entry of fs.readdirSync(src)) {
-      const srcPath = path.join(src, entry);
-      const destPath = path.join(dest, entry);
-      copyRecursiveSync(srcPath, destPath);
-    }
-  } else {
-    fs.mkdirSync(path.dirname(dest), { recursive: true });
-    fs.copyFileSync(src, dest);
-  }
-}
 
 function copyFile(src) {
   const rel = path.relative(appDir, src);
